Add s command that is similar to find + show
This makes it easy to use show without having to remember the
exact entry name.  It always prints the matching entry name so
you know which entry it is showing, and can cancel the display if
it doesn't match.  If the pattern matches multiple entries, it
shows them all so you can hit up, modify the pattern, and rerun.